+++ rules-engine/src/test/kotlin/com/wingedsheep/engine/scenarios/GrowingRitesOfItlimocScenarioTest.kt
@@ -4,7 +4,7 @@ import com.wingedsheep.engine.core.SelectCardsDecision
 import com.wingedsheep.engine.state.components.identity.CardComponent
 import com.wingedsheep.engine.support.GameTestDriver
 import com.wingedsheep.engine.support.TestCards
-import com.wingedsheep.mtg.sets.definitions.lci.cards.GrowingRitesOfItlimoc
+import com.wingedsheep.mtg.sets.definitions.xln.cards.GrowingRitesOfItlimoc
 import com.wingedsheep.sdk.core.Color
 import com.wingedsheep.sdk.core.Step
 import com.wingedsheep.sdk.model.Deck
