+++ rules-engine/src/test/kotlin/com/wingedsheep/engine/scenarios/TreasureMapScenarioTest.kt
@@ -6,7 +6,7 @@ import com.wingedsheep.engine.state.components.identity.CardComponent
 import com.wingedsheep.engine.state.components.identity.DoubleFacedComponent
 import com.wingedsheep.engine.support.GameTestDriver
 import com.wingedsheep.engine.support.TestCards
-import com.wingedsheep.mtg.sets.definitions.lci.cards.TreasureMap
+import com.wingedsheep.mtg.sets.definitions.xln.cards.TreasureMap
 import com.wingedsheep.mtg.sets.tokens.PredefinedTokens
 import com.wingedsheep.sdk.core.Color
 import com.wingedsheep.sdk.core.CounterType
